fix: ajuste entity

+++ src/main/java/com/faria/vehicle_allocation_system/service/ClienteService.java
@@ -6,37 +6,51 @@ import com.faria.vehicle_allocation_system.mapper.ClienteMapper;
 import com.faria.vehicle_allocation_system.repository.ClienteRepository;
 import com.faria.vehicle_allocation_system.util.ValidadorIdade;
 import jakarta.persistence.EntityNotFoundException;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class ClienteService {
 
     private final ClienteRepository clienteRepository;
-    private final ClienteMapper clienteMapper;
     private final ValidadorIdade validadorIdade;
 
-    public ClienteDTO buscarCliente(UUID clienteId) {
-        return clienteRepository.findById(clienteId)
-                .map(clienteMapper::toDTO)
-                .orElseThrow(() -> new EntityNotFoundException("Cliente com ID " + clienteId + " não encontrado."));
+    public ClienteService(ClienteRepository clienteRepository, ValidadorIdade validadorIdade) {
+        this.clienteRepository = clienteRepository;
+        this.validadorIdade = validadorIdade;
+    }
+
+    public ClienteDTO buscarCliente(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("O ID do cliente não pode ser nulo.");
+        }
+
+        ClienteEntity clienteEntity = clienteRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Cliente com ID " + id + " não encontrado."));
+
+        return ClienteMapper.INSTANCE.toDTO(clienteEntity); // MapStruct converte automaticamente
     }
 
     public ClienteDTO cadastrarCliente(ClienteDTO clienteDTO) {
         validadorIdade.validarIdade(clienteDTO.getDataNascimento());
 
-        ClienteEntity clienteEntity = clienteMapper.toEntity(clienteDTO);
+        ClienteEntity clienteEntity = ClienteMapper.INSTANCE.toEntity(clienteDTO);
         ClienteEntity savedEntity = clienteRepository.save(clienteEntity);
 
-        return clienteMapper.toDTO(savedEntity);
+        return ClienteMapper.INSTANCE.toDTO(savedEntity);
     }
 
-    public void deletarCliente(UUID clienteId) {
-        ClienteEntity clienteExistente = clienteRepository.findById(clienteId)
-                .orElseThrow(() -> new RuntimeException("Cliente não encontrado."));
+    public void deletarCliente(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("O ID do cliente não pode ser nulo.");
+        }
+
+        ClienteEntity clienteExistente = clienteRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Cliente não encontrado com ID: " + id));
+
         clienteRepository.delete(clienteExistente);
     }
 }
